impute tAML transfers back to tMDS for 2010-2012
Change of strategy. Keep tAML and tMDS as is up to 2009. Use sex-specific linear fits to tAML cases versus years 2001-2009 to predict tAML cases in 2010-2012 and then, using set.seed and sample, transfer a random grab of surplus cases to MDS, sex and year specifically. I'm upset that SEER stopped using 9987 for tMDS and instead mixed it in with tAML in 9920. This may be fine from a subsequent treatment perspective, but it stinks from a carcinogenesis perspective.